# Optimize method comparison