--- a/src/FakeItEasy/Core/WrappedObjectRule.cs
+++ b/src/FakeItEasy/Core/WrappedObjectRule.cs
@@ -1,8 +1,6 @@
 namespace FakeItEasy.Core
 {
     using System;
-    using System.Reflection;
-    using static FakeItEasy.ObjectMembers;
 
     /// <summary>
     /// A call rule that applies to any call and just delegates the