Connects to #1848.

I hesitate to say fix, but it addresses at least one area of slowness identified there.

Has this affect on runtimes (6.1.0 is the build before we changed a lot of method comparisons to see if a call had the same base method as a given method).

 Method |             Job |      Mean |     Error |    StdDev | Ratio | RatioSD |
------- |---------------- |----------:|----------:|----------:|------:|--------:|
 Equals |           6.1.0 |  4.738 us | 0.3711 us | 0.2208 us |  1.00 |    0.00 |
 Equals |           7.1.0 | 11.110 us | 0.6914 us | 0.4115 us |  2.35 |    0.14 |
 Equals | 7.1.1-alpha.0.5 |  4.340 us | 0.3255 us | 0.2153 us |  0.91 |    0.08 |

![image](https://user-images.githubusercontent.com/3275797/123670617-67a46900-d80b-11eb-9f92-3045c579a2d2.png)

and

 Method                 |             Job |     Mean |     Error |    StdDev | Ratio | RatioSD |
----------------------- |---------------- |---------:|----------:|----------:|------:|--------:|
 UnconfiguredBoolMethod |           6.1.0 | 5.421 us | 0.4837 us | 0.3199 us |  1.00 |    0.00 |
 UnconfiguredBoolMethod |           7.1.0 | 7.421 us | 0.3595 us | 0.2139 us |  1.38 |    0.10 |
 UnconfiguredBoolMethod | 7.1.1-alpha.0.5 | 4.998 us | 0.3071 us | 0.2032 us |  0.92 |    0.05 |
 
![image](https://user-images.githubusercontent.com/3275797/123670668-72f79480-d80b-11eb-89d9-5291d05e6dab.png)