@blairconrad 2021-06-29
Benchmarks were gathered by running 
```
.\benchmark.ps1 --filter=FakeItEasy.Benchmark.Program.SlowObjectBenchmark.UnconfiguredBoolMethod 6.1.0 7.1.0 7.1.1-alpha.0.5
.\benchmark.ps1 --filter=FakeItEasy.Benchmark.Program.SlowObjectBenchmark.Equals 6.1.0 7.1.0 7.1.1-alpha.0.5
```
(mostly because I didn't want to walk away from the computer to run all of them at once) on 
https://github.com/blairconrad/FakeItEasy.Benchmark/tree/f52025ce8cf341a7f25950c84eadcbb15adb302b

The benchmarks project is very rough, more or less the least I could do to get bars to come out. Were we to formalize the project, it may make sense to replace the plotting library with something .NETish. Maybe https://github.com/ScottPlot/ScottPlot, which I've not tried, but it looks like a good fit.